Split bottomtabnav into several smaler files and fixed bug with nav to SS
