+++ navigation/BottomTabNavigator.tsx
@@ -5,46 +5,33 @@
 
 import { Ionicons, MaterialIcons  } from '@expo/vector-icons';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { createStackNavigator } from '@react-navigation/stack';
 import * as React from 'react';
 
 import Colors from '../constants/Colors';
 import useColorScheme from '../hooks/useColorScheme';
 
-import CompaniesScreen from '../screens/CompaniesScreen';
-import CompanyDetailsScreen from '../screens/CompanyDetailsScreen';
-import EventDetailsScreen from '../screens/EventDetailsScreen';
-import EventListScreen from "../screens/EventListScreen";
-import SSsCompaniesScreen from '../screens/studentSessions/SSsCompaniesScreen';
-import SSsListScreen from "../screens/studentSessions/SSsListScreen";
-import MapScreen from "../screens/MapScreen";
-import ProfileScreen from '../screens/ProfileScreen';
-import TicketsScreen from '../screens/TicketsScreen';
-import QRScreen from '../screens/QRScreen';
-import ZoomMapScreen from '../screens/ZoomMapScreen';
-import { Map } from '../components/maps/MapProps';
-import EditProfileScreen from '../screens/EditProfileScreen';
 import { getMe, Role, User } from '../api/users';
 
-import { Platform } from 'react-native';
-import SSsDetailsScreen from '../screens/studentSessions/SSsDetailsScreen';
-import SSsApplicationScreen from '../screens/studentSessions/SSsApplicationSreen';
-import SSsApplicationsListScreen from '../screens/studentSessions/SSsApplicationsListScreen';
-import { Text, View } from '../components/Themed';
+import { View } from '../components/Themed';
 import { API } from '../api';
-import { SSApplication } from '../api/sSApplications';
-import SSsApplicationDetailsScreen from '../screens/studentSessions/SSsApplicationDetailsScreen';
 import { LogoutButton } from '../components/profileScreen/Buttons';
 import ScreenActivityIndicator from '../components/ScreenActivityIndicator';
 import { useContext } from 'react';
 import { AuthContext } from '../components/AuthContext';
+import { MapNavigator } from './MapNavigator';
+import { ProfileNavigator } from './ProfileNavigator';
+import { EventsNavigator } from './EventsNavigator';
+import { CompaniesNavigator } from './CompaniesNavigator';
+import { SSsStudentNavigator } from './SSsStudentNavigator';
+import { SSsCRepNavigator } from './SSsCRepNavigator';
 
 
 export type BottomTabParamList = {
   Companies: undefined;
   Maps: undefined;
   Profile: undefined;
-  SSs: undefined;
+  SSsStudent: undefined;
+  SSsCRep: {companyId: number;};
   Events: undefined
 };
 
@@ -73,7 +60,7 @@ export default function BottomTabNavigator() {
     authContext.signOut();
   };
 
-  if(isLoading || user == null) {
+  if(isLoading || !user) {
     return (
     <View>
       <ScreenActivityIndicator />
@@ -100,22 +87,25 @@ export default function BottomTabNavigator() {
         component={EventsNavigator}
         options={{ tabBarIcon: ({ color }) => <TabBarMaterialIcon name="event" color={color} />, }}
       />
+      {user.role !== Role.CompanyRepresentative ? 
       <BottomTab.Screen
-        name="SSs"
-        component={SSsNavigator}
+        name="SSsStudent"
+        component={SSsStudentNavigator}
         options={{
           title: 'Student Sessions',
           tabBarIcon: ({ color }) => <TabBarMaterialIcon name="forum" color={color} />,
         }}
-        listeners={({ navigation }) => ({
-          tabPress: (e) => {
-            if (navigation.canGoBack()) {
-              //navigation.popToTop()
-            }
-            user.role === Role.CompanyRepresentative ? navigation.replace('SSsListScreen', { companyId }) : navigation.replace('SSsCompaniesScreen')
-          },
-        })}
+      /> : companyId &&
+      <BottomTab.Screen
+        name="SSsCRep"
+        component={SSsCRepNavigator}
+        options={{
+          title: 'Student Sessions',
+          tabBarIcon: ({ color }) => <TabBarMaterialIcon name="forum" color={color} />,
+        }}
+        initialParams={{companyId: companyId}}
       />
+      }
       <BottomTab.Screen
         name="Profile"
         component={ProfileNavigator}
@@ -135,170 +125,4 @@ function TabBarMaterialIcon(props: { name: React.ComponentProps<typeof MaterialI
   return <MaterialIcons size={30} style={{ marginBottom: -3 }} {...props} />;
 }
 
-export type CompanyStackParamList = {
-  CompaniesScreen: undefined;
-  CompanyDetailsScreen: {
-    id: number;
-  };
-};
-const CompanyStack = createStackNavigator<CompanyStackParamList>();
-function CompaniesNavigator() {
-  return (
-    <CompanyStack.Navigator>
-      <CompanyStack.Screen
-        name="CompaniesScreen"
-        component={CompaniesScreen}
-        options={{ title: 'Companies', headerTitle: 'Companies' }}
-      />
-      <CompanyStack.Screen
-        name="CompanyDetailsScreen"
-        component={CompanyDetailsScreen}
-        options={{ title: 'Company Details', headerTitle: 'Company Details' }}
-      />
-    </CompanyStack.Navigator>
-  );
-}
 
-export type EventStackParamlist = {
-  EventListScreen: undefined;
-  EventDetailsScreen: {
-    id: number;
-  }
-}
-const EventStack = createStackNavigator<EventStackParamlist>();
-function EventsNavigator() {
-  return (
-    <EventStack.Navigator>
-      <EventStack.Screen
-        name="EventListScreen"
-        component={EventListScreen}
-        options={{ title: 'Events', headerTitle: 'Events' }}
-      />
-      <EventStack.Screen
-        name="EventDetailsScreen"
-        component={EventDetailsScreen}
-        options={{ title: 'Event Details', headerTitle: 'Event Details' }}
-      />
-    </EventStack.Navigator>
-  );
-}
-
-export type SSsStackParamlist = {
-  SSsCompaniesScreen: undefined;
-  SSsListScreen: {
-    companyId: number;
-  }
-  SSsDetailsScreen: {
-    companyId: number;
-    timeslotId: number;
-  }
-  SSsApplicationScreen: {
-    companyId: number;
-  }
-  SSsApplicationsListScreen: undefined;
-  SSsApplicationDetailsScreen: {
-    applicationId: number;
-  }
-}
-
-const SSsStack = createStackNavigator<SSsStackParamlist>();
-function SSsNavigator() {
-  return (
-    <SSsStack.Navigator>
-      <SSsStack.Screen
-        name="SSsCompaniesScreen"
-        component={SSsCompaniesScreen}
-        options={{ title: 'Student Sessions', headerTitle: 'Student Sessions' }}
-      />
-      <SSsStack.Screen
-        name="SSsListScreen"
-        component={SSsListScreen}
-        options={{ title: 'Student Sessions List', headerTitle: 'Student Sessions List' }}
-      />
-      <SSsStack.Screen
-        name="SSsDetailsScreen"
-        component={SSsDetailsScreen}
-        options={{ title: 'Session Details', headerTitle: 'Session Details' }}
-      />
-      <SSsStack.Screen
-        name="SSsApplicationScreen"
-        component={SSsApplicationScreen}
-        options={{ title: 'Application', headerTitle: 'Application' }}
-      />
-      <SSsStack.Screen
-        name="SSsApplicationsListScreen"
-        component={SSsApplicationsListScreen}
-        options={{ title: 'Applications', headerTitle: 'Applications' }}
-      />
-      <SSsStack.Screen
-        name="SSsApplicationDetailsScreen"
-        component={SSsApplicationDetailsScreen}
-        options={{ title: 'Application Details', headerTitle: 'Application Details' }}
-      />
-    </SSsStack.Navigator>
-  );
-}
-
-export type MapStackParamList = {
-  MapScreen: undefined;
-  ZoomMapScreen: {
-    map: Map;
-  }
-}
-const MapStack = createStackNavigator<MapStackParamList>();
-function MapNavigator() {
-  return (
-    <MapStack.Navigator>
-      <MapStack.Screen
-        name="MapScreen"
-        component={MapScreen}
-        options={{ title: 'Maps', headerTitle: 'Maps' }} />
-    <MapStack.Screen
-        name="ZoomMapScreen"
-        component={ZoomMapScreen}
-        options={{ title: 'Map', headerTitle: 'Map' }} />
-    </MapStack.Navigator>
-  )
-}
-
-export type ProfileStackParamList = {
-  ProfileScreen: undefined;
-  EditProfileScreen: undefined;
-  EventDetailsScreen: {
-    id: number;
-  },
-  TicketsScreen: undefined;
-  QRScreen: undefined,
-}
-const ProfileStack = createStackNavigator<ProfileStackParamList>();
-function ProfileNavigator() {
-  return (
-    <ProfileStack.Navigator>
-      <ProfileStack.Screen
-        name="ProfileScreen"
-        component={ProfileScreen}
-        options={{ title: 'Profile', headerTitle: 'Profile' }}
-      />
-      <ProfileStack.Screen
-        name="EditProfileScreen"
-        component={EditProfileScreen}
-        options={{ title: 'Edit Profile', headerTitle: 'Edit Profile' }}
-      />
-      <ProfileStack.Screen 
-        name="EventDetailsScreen"
-        component={EventDetailsScreen}
-        options={{ headerTitle: 'Event' }}
-      />
-      <ProfileStack.Screen 
-        name="TicketsScreen"
-        component={TicketsScreen}
-        options={{ title: 'Tickets', headerTitle: 'Tickets' }}
-      />
-      <ProfileStack.Screen
-        name="QRScreen"
-        component={QRScreen}
-        options={{ headerTitle: 'QR' }}
-      />
-    </ProfileStack.Navigator>
-  );
-}
